Move scap_bpf.* to userspace/libscap/engine/bpf/
